requirements.txt
requests2
aiohttp